% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbc2dbf.R
\name{dbc2dbf}
\alias{dbc2dbf}
\title{Decompress a .dbc file into a .dbf}
\usage{
dbc2dbf(input.file, output.file)
}
\arguments{
\item{input.file}{The name of the .dbc file (including extension)}

\item{output.file}{The output file name (including extension)}
}
\description{
This function allows you decompress a .dbc file into its .dbf counterpart.
}
\examples{
dbc2dbf("mydata.dbc","mydata.dbf")
}
\keyword{dbc}
\keyword{dbf}

