% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbf2dbc.R
\name{dbf2dbc}
\alias{dbf2dbc}
\title{Compress a standard DBF file into a DBC file (Experimental)}
\usage{
dbf2dbc(input, output)
}
\arguments{
\item{input}{The input DBF file path.}

\item{output}{The output DBC file path.}
}
\value{
TRUE if successful.
}
\description{
This function compresses a standard DBF file into a DBC file using the PKWare DCL Implode algorithm. This functionality is experimental and may be subject to change.
}
\examples{
\dontrun{
# Compress a file
dbf2dbc("data.dbf", "data.dbc")
}
}
