% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dbc.R
\name{read.dbc}
\alias{read.dbc}
\title{Read a .dbc File}
\usage{
read.dbc(filename, keep.dbf = FALSE, ...)
}
\arguments{
\item{filename}{The name of the .dbc file (including extension)}

\item{keep.dbf}{Keeps the temporary .dbf file instead of deleting it (Defaults to FALSE)}

\item{...}{Further arguments to be passed to read.dbf}
}
\description{
This function allows you to read a .dbc (compressed .dbf) file into a data frame.
}
\examples{
df <- read.dbc("mydata.dbc")
}
\keyword{dbc}

