% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dbc.R
\name{read.dbc}
\alias{read.dbc}
\title{Read Data Stored in DBC (Compressed DBF) Files}
\usage{
read.dbc(file, ...)
}
\arguments{
\item{file}{The name of the DBC file (including extension)}

\item{...}{Further arguments to be passed to \code{\link[foreign]{read.dbf}}}
}
\value{
A data.frame of the data from the DBC file.
}
\description{
This function allows you to read a DBC (compressed DBF) file into a data frame.
}
\details{
DBC is the extension for compressed DBF files (from the 'XBASE' family of databases).
This is a proprietary file format used by the Brazilian government to publish public healthcare data, and it is not related to the FoxPro or CANdb DBC file formats.

The \code{read.dbc} function will decompress the input DBC file into a temporary DBF file and call \code{\link[foreign]{read.dbf}} from the \code{foreign} package to read it into a data frame.
}
\note{
DATASUS is the name of the Department of Informatics of the Brazilian Health System (Sistema Único de Saúde - SUS) and is responsible for publishing public healthcare data in Brazil.
Besides the DATASUS, the Brazilian National Agency for Supplementary Health (ANS) also uses this file format for its public data.

This function was tested using files from both DATASUS and ANS to ensure compliance with the format, and hence ensure its usability by researchers.

Neither this project, nor its author, has any association with the Brazilian government.
}
\examples{
# The 'sids.dbc' file is the compressed version of 'sids.dbf' from the "foreign" package.
file <- system.file("files/sids.dbc", package="read.dbc")
sids <- read.dbc(file)
str(sids)
summary(sids)

# This is a small subset of U.S. NOAA storm database.
file <- system.file("files/storm.dbc", package="read.dbc")
storm <- read.dbc(file)
head(storm)
str(storm)

}
\seealso{
\code{\link{dbc2dbf}}
}
\author{
Daniela Petruzalek, \email{daniela.petruzalek@gmail.com}
}
\keyword{datasus}
\keyword{dbc}
